Fix pallet-commitments compilation

+++ pallets/commitments/src/mock.rs
@@ -109,7 +109,7 @@ impl pallet_commitments::GetTempoInterface for MockTempoInterface {
     fn get_epoch_index(netuid: NetUid, cur_block: u64) -> u64 {
         let tempo = 360; // TODO: configure SubtensorModule in this mock
         let tempo_plus_one: u64 = tempo.saturating_add(1);
-        let netuid_plus_one: u64 = (netuid as u64).saturating_add(1);
+        let netuid_plus_one: u64 = (u16::from(netuid) as u64).saturating_add(1);
         let block_with_offset: u64 = cur_block.saturating_add(netuid_plus_one);
 
         block_with_offset.checked_div(tempo_plus_one).unwrap_or(0)
